+++ config/local/hmri_local_defaults.m
@@ -233,7 +233,7 @@ hmri_def.UNICORT.MT = false;
 %--------------------------------------------------------------------------
 hmri_def.PDproc.calibr    = 1;   % Calibration of the PD map (if PDw, T1w, 
     % B1map available and RF sensitivity bias correction applied somehow)
-    % based on PD(WM) = 69% [Tofts 2003]. 
+hmri_def.PDproc.WMval = 69;  % WM percentage based on PD(WM) = 69% [Tofts 2003].   
 hmri_def.PDproc.WBMaskTh = 0.1;  % Threshold for calculation of whole-brain mask from TPMs
 hmri_def.PDproc.WMMaskTh = 0.95; % Threshold for calculation of white-matter mask from TPMs
 hmri_def.PDproc.biasreg  = 10^(-5);